+++ NEWS.rst
@@ -5,6 +5,7 @@ New features
 ------------
 * #40619: Added substitute to associations tab under organisations
 * #40828: Remove `default_page_size` parameter, making paged endpoints return the full sized result per default
+* #40828: Remove `tree_search_limit` parameter
 
 Version 1.12.0, 2021-01-08
 ==========================